+++ projeto-biblioteca/Repositories/interfaces/IReservationRepository.cs
@@ -2,7 +2,7 @@ using pBiblioteca.Models;
 
 public interface IReservationRepository
 {
-    TbReservation? GetById(long id);
+    TbReservation? GetById(int id);
 
     /// Retorna uma reserva ativa (Status = true) do usuário para um livro.
     TbReservation? GetActiveReservationForUserAndBook(string cpf, string isbn);
@@ -11,7 +11,7 @@ public interface IReservationRepository
     List<TbReservation> GetActiveQueueByBook(string isbn);
 
     /// Quantas reservas ativas existem antes desta reserva na fila do livro.
-    int CountActiveBefore(string isbn, DateOnly reservationDate, long reservationId);
+    int CountActiveBefore(string isbn, DateOnly reservationDate, int reservationId);
 
     /// Cria a reserva e registra o vínculo com o livro.
     void CreateReservationWithBook(TbReservation reservation, string isbn);
